Replace modal form with EmailContactForm component
Deleted the `modalOpen.js` file and replaced it with the new `EmailContactForm.js` component that was created. The `Feedback.js` file was updated to use this new EmailContactForm component, and a emailjs library dependency was added to enable email sending functionality. The `Feedback.css` and `Feedback.scss` files were updated to properly style the new form component.